backend/api/serializers.py
from rest_framework import serializers
from .models import StudentModel


class StudentSerializer(serializers.ModelSerializer):
    class Meta:
        model = StudentModel
        fields = ['id', 'student_name', 'student_email']
