backend/requirements.txt
django == 5.0.4
djangorestframework == 3.15.1
flake8 == 7.0.0
django-cors-headers==4.3.1

